Fixes for Visual Studio
The main fixes are for the Windows multiprocessor code which now uses
std::unique_ptr to wrap dynamic allocation of an array to hold the
thread ids rather than a variable length array which is a GNU extension
to C++.

A couple of source files used 'or' and 'and' instead of || and &&; these
have also been replaced.

Replacement code for sys/time.h has been added to clearcut.cpp (not a
standard header on Windows).

Now that CMake is used, the environment variable to indicate a UNIX-like
OS is simply UNIX, and this has been used in place of the more
complicated OR statement.